Changed manual test data to better reflect it on the current database style in a 'P3000.py' file.
Added a new goal in 'TODO' section of a 'README.md' file.

'bot_functions.py':
Removed an unused import 'MonthsButton'.
Updated comments in a 'birthday_btn' function.
Changed a way of interaction with a 'step' variable within the function.

'database_functions.py':
Corrected a typo in a description of a 'database_write' function.
Removed new lines when creating a new row in a 'database_write' function.
Removed an unused comment in a 'database_remove' function.

'session_functions.py':
Changed a starting data in a 'session_start' function.
Changed a file data interaction in a 'session_user_data_extract' function.
Changed a file data interaction in a 'session_user_data_write' function.

'text_responses.py':
Added a comment to explain what is a new lone at the top of the file.
Overall updated text responses.

I believe that the bot is ready for release!
Known issues: spam in logs.